--- backend/src/main/java/com/woori/wonfit/member/membersearchlog/dto/MemberSearchLogResponse.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.membersearchlog.dto;
-
-public class MemberSearchLogResponse {
-}
