--- backend/src/main/java/com/woori/wonfit/member/membersearchlog/service/MemberSearchLogService.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.member.membersearchlog.service;
-
-public class MemberSearchLogService {
-}
