+++ backend/src/main/java/com/woori/wonfit/member/searchlog/dto/MemberSearchLogResponse.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.member.searchlog.dto;
+
+public class MemberSearchLogResponse {
+}
