+++ backend/src/main/java/com/woori/wonfit/member/searchlog/service/MemberSearchLogService.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.member.searchlog.service;
+
+public class MemberSearchLogService {
+}
