+++ backend/src/main/java/com/woori/wonfit/product/deposit/dto/ProductDepositResponse.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.product.deposit.dto;
+
+public class ProductDepositResponse {
+}
