+++ backend/src/main/java/com/woori/wonfit/product/deposit/service/DepositService.java
@@ -1,0 +1,4 @@
+package com.woori.wonfit.product.deposit.service;
+
+public class DepositService {
+}
