--- backend/src/main/java/com/woori/wonfit/product/productdeposit/dto/ProductDepositResponse.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.product.productdeposit.dto;
-
-public class ProductDepositResponse {
-}
