--- backend/src/main/java/com/woori/wonfit/product/productdeposit/service/DepositService.java
@@ -1,4 +1,0 @@
-package com.woori.wonfit.product.productdeposit.service;
-
-public class DepositService {
-}
